Restructure swap setting classes
