+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/swapxxx/ISwapQuote.kt
@@ -1,12 +1,13 @@
 package io.horizontalsystems.bankwallet.modules.swapxxx
 
 import io.horizontalsystems.bankwallet.modules.send.SendModule
+import io.horizontalsystems.bankwallet.modules.swapxxx.settings.ISwapSettingField
 import io.horizontalsystems.bankwallet.modules.swapxxx.ui.SwapDataField
 import io.horizontalsystems.marketkit.models.BlockchainType
 import java.math.BigDecimal
 
 interface ISwapQuote {
-    fun getSettingFields() : List<SwapSettingField>
+    fun getSettingFields() : List<ISwapSettingField>
 
     val amountOut: BigDecimal
     val fields: List<SwapDataField>
@@ -17,7 +18,7 @@ class SwapQuoteUniswap(
     override val amountOut: BigDecimal,
     override val fields: List<SwapDataField>,
     override val fee: SendModule.AmountData?,
-    private val settingFields: List<SwapSettingField>,
+    private val settingFields: List<ISwapSettingField>,
 ) : ISwapQuote {
     override fun getSettingFields() = settingFields
 }
@@ -28,7 +29,7 @@ class SwapQuoteUniswapV3(
     override val fee: SendModule.AmountData?,
     private val blockchainType: BlockchainType,
 ) : ISwapQuote {
-    override fun getSettingFields(): List<SwapSettingField> {
+    override fun getSettingFields(): List<ISwapSettingField> {
         TODO()
 //        return listOf(SwapSettingFieldRecipient(blockchainType))
     }
@@ -40,7 +41,7 @@ class SwapQuoteOneInch(
     override val fee: SendModule.AmountData?,
     private val blockchainType: BlockchainType,
 ) : ISwapQuote {
-    override fun getSettingFields(): List<SwapSettingField> {
+    override fun getSettingFields(): List<ISwapSettingField> {
         TODO()
 //        return listOf(SwapSettingFieldRecipient(blockchainType))
     }
