+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/swapxxx/SwapViewModel.kt
@@ -8,6 +8,7 @@ import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewModelScope
 import io.horizontalsystems.bankwallet.core.App
 import io.horizontalsystems.bankwallet.modules.swapxxx.providers.ISwapXxxProvider
+import io.horizontalsystems.bankwallet.modules.swapxxx.settings.SwapSettingsService
 import io.horizontalsystems.marketkit.models.Token
 import kotlinx.coroutines.launch
 import java.math.BigDecimal
