+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/swapxxx/settings/ISwapSettingField.kt
@@ -1,0 +1,15 @@
+package io.horizontalsystems.bankwallet.modules.swapxxx.settings
+
+import androidx.compose.runtime.Composable
+import androidx.navigation.NavController
+
+interface ISwapSettingField {
+    val id: String
+
+    @Composable
+    fun GetContent(
+        navController: NavController,
+        onError: (Throwable?) -> Unit,
+        onValueChange: (Any?) -> Unit
+    )
+}
