+++ app/src/main/res/navigation/main_graph.xml
@@ -332,8 +332,8 @@
         android:id="@+id/swapSelectProvider"
         android:name="io.horizontalsystems.bankwallet.modules.swapxxx.SwapSelectProviderFragment" />
     <fragment
-        android:id="@+id/swapProviderSettings"
-        android:name="io.horizontalsystems.bankwallet.modules.swapxxx.SwapProviderSettingsFragment" />
+        android:id="@+id/swapSettings"
+        android:name="io.horizontalsystems.bankwallet.modules.swapxxx.settings.SwapSettingsFragment" />
     <fragment
         android:id="@+id/swapConfirm"
         android:name="io.horizontalsystems.bankwallet.modules.swapxxx.SwapConfirmFragment" />
